@@ -3,23 +3,26 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "axios": "^0.18.0",
-    "bufferutil": "^4.0.1",
+    "@mdi/font": "^5.8.55",
+    "@mdi/js": "^5.8.55",
+    "@mdi/react": "^1.4.0",
+    "axios": "^0.21.0",
+    "bufferutil": "^4.0.2",
     "canvas": "^2.6.1",
-    "change-case": "^4.1.1",
+    "change-case": "^4.1.2",
     "classnames": "^2.2.6",
-    "jwt-decode": "^2.2.0",
-    "mdi-react": "^7.3.0",
-    "node-sass": "^4.14.1",
-    "react": "^16.6.3",
-    "react-dom": "^16.6.3",
-    "react-redux": "^5.1.1",
-    "react-router-dom": "^4.3.1",
-    "react-scripts": "2.1.1",
-    "redux": "^4.0.1",
+    "date-fns": "^2.16.1",
+    "jwt-decode": "^3.1.2",
+    "react": "^17.0.1",
+    "react-dom": "^17.0.1",
+    "react-redux": "^7.2.2",
+    "react-router-dom": "^5.2.0",
+    "react-scripts": "^4.0.1",
+    "redux": "^4.0.5",
+    "redux-promise": "^0.6.0",
     "redux-thunk": "^2.3.0",
-    "request": "^2.88.2",
-    "utf-8-validate": "^5.0.2"
+    "sass": "^1.30.0",
+    "utf-8-validate": "^5.0.3"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -27,7 +30,7 @@
     "test": "react-scripts test",
     "eject": "react-scripts eject"
   },
-  "proxy": "http://localhost:5000",
+  "proxy": "http://localhost:8802",
   "eslintConfig": {
     "extends": "react-app"
   },
