@@ -16,23 +16,24 @@ class Login extends Component {
   }
 
   componentDidMount() {
-    // If logged in and user navigates to Login page, should redirect them to dashboard
-    if (this.props.auth.isAuthenticated) {
-      this.props.history.push("/dashboard");
-    }
+      // If logged in and user navigates to Login page, should redirect them to dashboard
+      if (this.props.auth.isAuthenticated) {
+        this.props.history.push("/dashboard");
+      }
   }
 
-  // npx react-codemod rename-unsafe-lifecycles
-  UNSAFE_componentWillReceiveProps(nextProps) {
-    if (nextProps.auth.isAuthenticated) {
-      this.props.history.push("/dashboard"); // push user to dashboard when they login
-    }
+  static getDerivedStateFromProps(nextProps) {
+      if (nextProps.auth.isAuthenticated) {
+          nextProps.history.push("/dashboard"); // push user to dashboard when they login
+      }
 
-    if (nextProps.errors) {
-      this.setState({
-        errors: nextProps.errors
-      });
-    }
+      if (nextProps.errors) {
+          return ({
+              errors: nextProps.errors
+          });
+      }
+
+      return null;
   }
 
   onChange = e => {
@@ -69,6 +70,7 @@ class Login extends Component {
               error={errors.email}
               id="email"
               type="email"
+              autoComplete="username"
               className={classnames("", { invalid: errors.email || errors.emailnotfound })} />
             <span className="input-field-error-text">
               {errors.email}
