@@ -1,8 +1,18 @@
 import React, { Component } from "react";
-import { Link } from "react-router-dom";
+import { Link, withRouter } from "react-router-dom";
 import PropTypes from "prop-types";
+import Icon from '@mdi/react';
+import { mdiViewDashboardVariantOutline, mdiBarcode, mdiChartGantt, mdiCommentTextMultipleOutline } from '@mdi/js';
 
-export default class ButtonSelector extends Component {
+class ButtonSelector extends Component {
+    shouldComponentUpdate(nextProps, nextState) {
+        if (nextProps.location.pathname.slice(1) !== nextProps.buttonSelectorSelectedOption) {
+            this.handleClick(nextProps.location.pathname.slice(1));
+        }
+
+        return true;
+    }
+
     handleClick(value) {
         this.props.handleClick(value);
     }
@@ -11,27 +21,48 @@ export default class ButtonSelector extends Component {
         return (
           <div className="button-selector">
             <Link
-              className={this.props.selectedOption === "talk-ratio" ? "button-selector-item active" : "button-selector-item"}
+              className={this.props.buttonSelectorSelectedOption === "dashboard" ? "button-selector-item active" : "button-selector-item"}
+              data-attr-name="dashboard"
+              to="/dashboard"
+              onClick={this.handleClick.bind(this, "dashboard")}>
+              <Icon path={mdiViewDashboardVariantOutline} className="button-selector-item-icon" size={1} />
+              <span className="button-selector-item-text">
+                Dashboard
+              </span>
+            </Link>
+
+            <Link
+              className={this.props.buttonSelectorSelectedOption === "talk-ratio" ? "button-selector-item active" : "button-selector-item"}
               data-attr-name="talk-ratio"
-              to="/dashboard/talk-ratio"
+              to="/talk-ratio"
               onClick={this.handleClick.bind(this, "talk-ratio")}>
-              Talk Ratio
+              <Icon path={mdiBarcode} className="button-selector-item-icon" size={1.5} />
+
+              <span className="button-selector-item-text">
+                Talk Ratio
+              </span>
             </Link>
 
             <Link
-              className={this.props.selectedOption === "turn-taking" ? "button-selector-item active" : "button-selector-item"}
+              className={this.props.buttonSelectorSelectedOption === "turn-taking" ? "button-selector-item active" : "button-selector-item"}
               data-attr-name="turn-taking"
-              to="/dashboard/turn-taking"
+              to="/turn-taking"
               onClick={this.handleClick.bind(this, "turn-taking")}>
-              Turn Taking
+              <Icon path={mdiChartGantt} className="button-selector-item-icon" size={1} />
+              <span className="button-selector-item-text">
+                Turn Taking
+              </span>
             </Link>
 
             <Link
-              className={this.props.selectedOption === "transcript" ? "button-selector-item active" : "button-selector-item"}
+              className={this.props.buttonSelectorSelectedOption === "transcript" ? "button-selector-item active" : "button-selector-item"}
               data-attr-name="transcript"
-              to="/dashboard/transcript"
+              to="/transcript"
               onClick={this.handleClick.bind(this, "transcript")}>
-              Transcript
+              <Icon path={mdiCommentTextMultipleOutline} className="button-selector-item-icon" size={1} />
+              <span className="button-selector-item-text">
+                Transcript
+              </span>
             </Link>
           </div>
         );
@@ -39,5 +70,7 @@ export default class ButtonSelector extends Component {
 }
 
 ButtonSelector.propTypes = {
-    selectedOption: PropTypes.string.isRequired
+    buttonSelectorSelectedOption: PropTypes.string.isRequired
 }
+
+export default withRouter(ButtonSelector);
