@@ -1,11 +1,8 @@
 import React, { Component } from 'react';
 
-import Parser from '../../../data/parser';
-
 import LegendButtonGroup from '../../legend/LegendButtonGroup';
 import TurnTakingSmall from '../turn-taking/TurnTakingSmall';
 import Script from '../transcript/Script';
-import displayLegendLabels from '../../legend/displayLegendLabels';
 
 import removeArrayValue from '../../../utils/removeArrayValue';
 
@@ -14,11 +11,14 @@ export default class Transcript extends Component {
         console.log("super Transcript");
         super(props);
 
-        // double width - for both left/right side of TurnTakingSmall chart
-        this.chartWidth = 2 * Parser.maxNTokens();
+        var parser = props.activeParser,
+            chartWidth = 2 * parser.maxNTokens(), // double width - for both left/right side of TurnTakingSmall chart
+            talkRatios = parser.talkRatios(),
+            transcript = parser.transcript();
 
         console.log("constructor");
         this.state = {
+            parser: parser,
             activeLabels: [],
             focusBox: {
                 topElId: 0,
@@ -26,11 +26,13 @@ export default class Transcript extends Component {
                 y: 0,
                 height: 0
             },
-            chartOffsetWidth: 0,
-            chartHeight: 0
+            chartWidth: chartWidth,
+            talkRatios: talkRatios,
+            transcript: transcript
         };
     }
 
+<<<<<<< HEAD
     componentDidMount() {
         // dynamically orient and size TurnTakingSmall chart
         var legendButtonGroup = document.getElementsByClassName("transcript-visualization-legend")[0],
@@ -48,6 +50,8 @@ export default class Transcript extends Component {
         });
     }
 
+=======
+>>>>>>> origin/main
     // same logic as in TurnTaking::handleFilterClick
     handleClick(label) {
         var activeLabels = this.state.activeLabels;
@@ -81,48 +85,53 @@ export default class Transcript extends Component {
         // focus the box
         turnTakingBarsSmall.scrollTo(0, topElId * this.barHeight);
 
+<<<<<<< HEAD
         console.log("Transcript::handleScroll");
         console.log("this", this);
+=======
+        var focusBox = {
+            topElId: topElId,
+            bottomElId: bottomElId,
+            height: boxHeight,
+        };
+
+>>>>>>> origin/main
         this.setState({
-            focusBox: {
-                topElId: topElId,
-                bottomElId: bottomElId,
-                height: boxHeight,
-            }
+            focusBox: focusBox
         });
     }
 
     render() {
       return (
-        <div className="transcript-visualization-container" style={{ marginLeft: this.state.chartOffsetWidth }}>
+        <div className="transcript-visualization-container">
           <div className="transcript-visualization-legend">
             <LegendButtonGroup
-              labels={displayLegendLabels({ type: "Teacher"})}
+              labels={this.state.parser.legendLabels({ type: "Teacher"})}
               displayRatio={true}
               activeLabels={this.state.activeLabels}
               handleClick={this.handleClick.bind(this)} />
             <LegendButtonGroup
-              labels={displayLegendLabels({ type: "Student"})}
+              labels={this.state.parser.legendLabels({ type: "Student"})}
               displayRatio={true}
               activeLabels={this.state.activeLabels}
               handleClick={this.handleClick.bind(this)} />
             <LegendButtonGroup
-              labels={displayLegendLabels({ type: "Technique"})}
+              labels={this.state.parser.legendLabels({ type: "Technique"})}
               displayRatio={true}
               activeLabels={this.state.activeLabels}
               handleClick={this.handleClick.bind(this)} />
           </div>
 
           <TurnTakingSmall
-            chartOffsetWidth={this.state.chartOffsetWidth}
-            chartWidth={this.chartWidth}
-            chartHeight={this.state.chartHeight}
+            parser={this.state.parser}
+            chartWidth={this.state.chartWidth}
+
             barHeight={this.barHeight}
             focusBox={this.state.focusBox} />
 
-          <div className="transcript-script-container" style={{ marginLeft: `${this.chartWidth}px` }}>
+          <div className="transcript-script-container" style={{ marginLeft: `${this.state.chartWidth}px` }}>
             <Script
-              data={Parser.transcript()}
+              transcript={this.state.transcript}
               activeLabels={this.state.activeLabels}
               focusBox={this.state.focusBox}
               handleScroll={this.handleScroll.bind(this)}
