@@ -1,8 +1,8 @@
 import defineInitialStyle from './defineInitialStyle';
 
-export default function(item, areBarsSmall = false) {
+export default function drawBarStyles(item, areBarsSmall = false) {
     var isStudentData = item.speakerPseudonym.includes("Student"),
-        isTeacherData = item.speakerPseudonym === "Teacher",
+        isTeacherData = !item.speakerPseudonym.includes("Student"),
         initialStyle = defineInitialStyle(item);
 
     var barWidth = item.nTokens,
