@@ -1,12 +1,17 @@
-import data from './data';
 import LegendLabels from '../fixtures/legend_labels';
 
-export default {
-    segments: data[0].data.segments,
+export default class Parser {
+    constructor(data) {
+        this.data = data;
+        this.segments = data.data.segments;
+    }
 
-    legendLabelValues: LegendLabels.map((item) => item.value),
+    legendLabels = function(options) {
+        var legendLabels = this.talkRatios().filter((item) => item.type === options.type);
+        return legendLabels;
+    }
 
-    transcript: function() {
+    transcript = function() {
         var transcript = [];
         var utteranceIndex = 0;
 
@@ -28,11 +33,13 @@ export default {
                                 (speakingTurn.speaker_pseudonym.includes("Class") ||
                                 speakingTurn.speaker_pseudonym.includes("Student")),
                             unclassifiedTeacherTalk = utterance.utterance_type.length === 0 &&
-                                speakingTurn.speaker_pseudonym.includes("Teacher"),
+                                !speakingTurn.speaker_pseudonym.includes("Student") && !speakingTurn.speaker_pseudonym.includes("Class") && !speakingTurn.speaker_pseudonym.includes("Video"),
+                            videoTeacherTalk = speakingTurn.speaker_pseudonym.includes("Video"),
+                            speakerPseudonym = speakingTurn.speaker_pseudonym.includes("Video") ? "Teacher (Video)" : speakingTurn.speaker_pseudonym,
                             dataRow = {
                                 id: utteranceIndex++,
                                 timestamp: [],
-                                speakerPseudonym: speakingTurn.speaker_pseudonym,
+                                speakerPseudonym: speakerPseudonym,
                                 speakerUtterances: [utterance.utterance],
                                 nTokens: utterance.n_tokens
                             };
@@ -45,6 +52,8 @@ export default {
                             dataRow = { ...dataRow, ...{ utteranceTypes: ["Assorted Student Talk"] } };
                         } else if (unclassifiedTeacherTalk) {
                             dataRow = { ...dataRow, ...{ utteranceTypes: ["Assorted Teacher Talk"] } };
+                        } else if (videoTeacherTalk) {
+                            dataRow = { ...dataRow, ...{ utteranceTypes: ["Video"] } };
                         } else {
                             dataRow = { ...dataRow, ...{ utteranceTypes: utterance.utterance_type } };
                         }
@@ -56,9 +65,9 @@ export default {
         });
 
         return transcript;
-    },
+    }
 
-    filteredTranscript: function(options) {
+    filteredTranscript = function(options) {
         var data = this.transcript();
         var activeFilters = options && options.activeFilters;
 
@@ -82,9 +91,9 @@ export default {
         }, []);
 
         return filteredTranscript;
-    },
+    }
 
-    drilldownTranscript: function(options) {
+    drilldownTranscript = function(options) {
         var data = this.transcript(),
             drilldownFilter = options && options.drilldownFilter;
 
@@ -106,9 +115,9 @@ export default {
         }, []);
 
         return drilldownTranscript;
-    },
+    }
 
-    expandedData: function(options) {
+    expandedData = function(options) {
         var activeFilters = options && options.activeFilters;
 
         var transcript = this.filteredTranscript({ activeFilters: activeFilters });
@@ -116,15 +125,15 @@ export default {
         return transcript.reduce((accumulator, turn, index, array) => {
             return accumulator.concat(turn.utterances);
         }, []);
-    },
-    maxNTokens: function(options) {
+    }
+    maxNTokens = function(options) {
         var activeFilters = options && options.activeFilters;
 
         var expandedData = this.expandedData({ activeFilters: activeFilters });
 
         return Math.max.apply(Math, expandedData.map((utterance) => utterance.nTokens));
-    },
-    collapsedData: function(options) {
+    }
+    collapsedData = function(options) {
         var activeFilters = options && options.activeFilters;
         var expandedData = this.expandedData({ activeFilters: activeFilters }),
             collapsedData = [];
@@ -151,9 +160,9 @@ export default {
         });
 
         return collapsedData;
-    },
+    }
 
-    parsedData: function(options) {
+    parsedData = function(options) {
         var activeFilters = options && options.activeFilters;
         var expandedData = this.expandedData({ activeFilters: activeFilters }),
             collapsedData = this.collapsedData({ activeFilters: activeFilters });
@@ -164,9 +173,9 @@ export default {
         };
 
         return parsedData;
-    },
+    }
 
-    talkRatios: function() {
+    talkRatios = function() {
         var expandedData = this.expandedData(), // get array of every utterance in the transcript
             legendLabels = LegendLabels,
             talkRatios = legendLabels.map((labelObj, index, array) => { // set up object to be returned
@@ -208,8 +217,23 @@ export default {
         });
 
         return talkRatios;
-    },
-    initializeSpeakerTotals: function() {
+    }
+
+    teacherTalkRatios = function() {
+        var talkRatios = this.talkRatios(),
+            teacherTalkRatios = talkRatios.filter((item) => item.type === "Teacher").reverse();
+
+        return teacherTalkRatios;
+    }
+
+    studentTalkRatios = function() {
+        var talkRatios = this.talkRatios(),
+            studentTalkRatios = talkRatios.filter((item) => item.type === "Student");
+
+        return studentTalkRatios;
+    }
+
+    initializeSpeakerTotals = function() {
         var legendLabels = LegendLabels;
         var speakerTotals = legendLabels.reduce((accumulator, labelObj, index, array) => {
             var speakerIsInArray = accumulator.filter((accumObj) => {
@@ -227,8 +251,8 @@ export default {
         }, []);
 
         return speakerTotals;
-    },
-    speakerTalkTotals: function() {
+    }
+    speakerTalkTotals = function() {
         var speakerTotals = this.initializeSpeakerTotals(),
             talkRatios = this.talkRatios();
 
