@@ -1,4 +1,4 @@
-export default [{
+var legendLabels = [{
     type: "Teacher",
     value: "Metacognitive Modeling Questions",
     text: "Metacognitive",
@@ -106,4 +106,14 @@ export default [{
     speakerType: "Teacher",
     textColor: "white",
     description: "This code is applied when a teacher asks a question that is managing behavior and/or task completion."
+}, {
+    type: "Teacher",
+    value: "Video",
+    text: "Video",
+    barColor: "#29CB97",
+    speakerType: "Teacher",
+    textColor: "white",
+    description: ""
 }];
+
+export default legendLabels;
