@@ -1,10 +1,11 @@
 import { createStore, applyMiddleware, compose } from "redux";
 import thunk from "redux-thunk";
+import promiseMiddleware from 'redux-promise';
 import rootReducer from "./reducers";
 
 const initialState = {};
 
-const middleware = [thunk];
+const middleware = [thunk, promiseMiddleware];
 
 const store = createStore(
   rootReducer,
