@@ -1,4 +1,4 @@
-export default function(decimalValue, precision, showValuesLessThanOne = true) {
+export default function formatPercentage(decimalValue, precision, showValuesLessThanOne = true) {
     var formattedValue = `${parseFloat(100 * decimalValue).toFixed(precision)}%`;
 
     if (showValuesLessThanOne && decimalValue > 0 && decimalValue < 0.01 && precision === 0) {
