@@ -1,3 +1,3 @@
-export default function(value) {
+export default function isObjectEmpty(value) {
     return Boolean(value && typeof value === 'object') && !Object.keys(value).length;
 }
