@@ -1,9 +1,9 @@
-export default function(value, array) {
+export default function removeArrayValue(value, array) {
     var index = array.indexOf(value);
 
     if (index > -1) {
         array.splice(index, 1);
     }
-    
+
     return array;
 }
