@@ -4,15 +4,14 @@ const bodyParser = require("body-parser");
 const passport = require("passport");
 
 const users = require("./routes/api/users");
+const datasets = require("./routes/api/datasets");
 
 const app = express();
 
+app.use(express.json({limit: '25mb'}));
+app.use(express.urlencoded({extended: true, limit: '25mb'}));
+
 // Bodyparser middleware
-app.use(
-  bodyParser.urlencoded({
-    extended: false
-  })
-);
 app.use(bodyParser.json());
 
 // DB Config
@@ -35,7 +34,8 @@ require("./config/passport")(passport);
 
 // Routes
 app.use("/api/users", users);
+app.use("/api/datasets", datasets);
 
-const port = process.env.PORT || 5000; // process.env.port is Heroku's port if you choose to deploy the app there
+const port = process.env.PORT || 8802; // process.env.port is Heroku's port if you choose to deploy the app there
 
 app.listen(port, () => console.log(`Server up and running on port ${port} !`));
